Make library search suffixes configurable via macro argument.
Kokkos team requested that lib/stubs, lib64/stubs library
search suffixes be added to the suffix list only when searching
for CUDA. This commit adds options to kokkos_find_imported and
kokkos_find_library to allow the library search suffixes to be
set by the caller. It also updates FindTPLCUDA to use this
feature to add the stub directories to the search list.